Add files via upload
The analysis was conducted using the Maven Pizza Challenge data set, and the Star schema was used to identify relationships between the fact and dimension tables. Custom columns and conditional columns were created using DAX functions, and a dashboard was developed based on the analysis. Additionally, row-level security was implemented.

The analysis yielded insights into the pizza restaurant's operations. For example, the analysis revealed the days and times when the restaurant is busiest. This information could help the restaurant plan its staffing and inventory levels more effectively. The analysis also provided information on how many pizzas are typically made during peak periods, which could be used to optimize the restaurant's production process.

Furthermore, the analysis identified the best and worst selling pizzas. This information could be used to adjust the restaurant's menu offerings and marketing strategies to better align with customer preferences. Finally, the analysis provided information on the restaurant's average order value, which could be used to optimize pricing strategies and maximize revenue.

Overall, the analysis provided valuable insights that could be used to improve the pizza restaurant's operations and profitability.